k-digital_2nd_스마트팩토리/papercup1/requirements.txt
alembic==1.14.1
blinker==1.8.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
filelock==3.16.1
Flask==3.0.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
fonttools==4.55.8
fsspec==2025.2.0
greenlet==3.1.1
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.5
kiwisolver==1.4.7
Mako==1.3.9
MarkupSafe==2.1.5
matplotlib==3.7.5
mpmath==1.3.0
mysql-connector-python==9.0.0
networkx==3.1
numpy==1.24.4
opencv-python==4.11.0.86
packaging==24.2
pandas==2.0.3
pillow==10.4.0
psutil==6.1.1
py-cpuinfo==9.0.0
PyAudio==0.2.14
pycparser==2.22
PyMySQL==1.1.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
scipy==1.10.1
seaborn==0.13.2
six==1.17.0
SQLAlchemy==2.0.38
sympy==1.13.3
torch==2.4.1
torchvision==0.19.1
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.72
ultralytics-thop==2.0.14
urllib3==2.2.3
Werkzeug==3.0.6
zipp==3.20.2
